Moved streamers controller
